# Reduce CString allocations in std as much as possible