… library/std/src/sys/wasi/fs.rs
… library/std/src/sys/wasi/os.rs
… library/std/src/sys_common/net.rs
… src/tools/miri/tests/fail/shims/fs/isolated_file.stderr