@hkratz 2022-02-05
Wouldn't it be better to have a `SmallCString` type, which works like a `SmallVec` and implements `core::ops::Deref` with `type Target = CStr`? Then we could have  `cstr_small()` returning `Result<SmallCString>` and just replace `cstr()` with `cstr_small()` whereever it make sense.