@SUPERCILEX 2022-02-27
@dtolnay Can I get some feedback on whether or not this is desired? I just completed similar work in nix: https://github.com/nix-rust/nix/pull/1656.