@SUPERCILEX 2022-02-28
Update: I've been analyzing memory allocations in one of my programs that does a lot of I/O and path allocations account for ~15% of *all* allocations in the program. That's huge for something that should be free!