@joshtriplett 2022-04-15
The concept of this seems reasonable to me. I agree with the suggestion to reduce the stack allocation to 512 bytes, and I do think the `From` impl would need to happen in a separate PR that needs an FCP. I also think the `run_with_cstr` approach may warrant an `inline` on run_with_cstr, and a check of the generated code to confirm that it doesn't generate an extra function call.