@SUPERCILEX 2022-04-16
Awesome! I created https://github.com/rust-lang/rust/pull/96104 to get that started.