@ivmarkov 2022-04-16
> Awesome! I created #96104 to get that started.

Reminder: as discussed earlier, on MCU targets like `espidf` we need the code path allocating 512 bytes on the stack either completely switched off, or operating with significantly smaller allocation. As in 16 to 32 bytes max. The average stack size on `espidf` is 3 to 4K.