@SUPERCILEX 2022-04-16
@ivmarkov Some random googling says the cache line is 32 bytes on an ESP32, so 32 bytes seems reasonable.