@SUPERCILEX 2022-04-20
Ok, I looked into the SmallCStr idea and it actually already exists: https://github.com/rust-lang/rust/blob/master/compiler/rustc_data_structures/src/small_c_str.rs. Unfortunately, that's in the compiler so we'd have to bring over all of SmallVec into std to make things work which doesn't seem like a good idea. Given those issues, I'll continue with the functional approach and figure what the assembly looks like. Should hopefully have time to do this before the end of the week.