@hkratz 2022-04-20
@SUPERCILEX I have been prototyping a minimal SmallCString implementation without external dependencies. It works fine except for some unnecessary copying. I will share the code in a draft PR shortly.