@SUPERCILEX 2022-04-20
As long as it's as efficient as single branch on "is bigger than X, then allocate, else use the stack" that works for me. One thing I just thought of: with the SmallCString approach, we're always going to use the same amount of stack space even if we don't need it. That probably doesn't matter, but I think the current approach allows decrementing %rsp only when we plan to use that space and avoid it if we're allocating.