@hkratz 2022-04-20
The reserved (stack) frame size is fixed for each function in Rust and the stack pointer is decremented accordingly in the beginning of the function (see e.g. [Godbolt](https://godbolt.org/z/xhnM1Gq6G)). So that would be the case for `run_with_cstr` if it is inlined as well.