@SUPERCILEX 2022-04-21
r? @joshtriplett 

@hkratz Actually, you inspired me to do the SmallCString thing since I realized we don't have to support mutability and can therefore have a much simpler implementation than SmallVec. Let me know if you have feedback on my implementation.

@joshtriplett 
- Why is the tool yelling at me to put stability attributes on `SmallCString` stuff? It's in a [private module](https://github.com/rust-lang/rust/blob/879aff385a5fe0af78f3d45fd2f0b8762934e41e/library/std/src/lib.rs#L569), so nothing should escape anyway. Also, did I put it in the right place? I had to make `common` public to one more level of crates.
- I went ahead and replaced CString usage for the platforms that were easy, but windows obviously sticks out as being a pain. If someone else wants to do it, feel free to cc me on their PR and I'll be happy to review, but I don't plan on doing anything myself since I don't really care about windows.
- Who else would want a special stack size? I'm pretty sure I set up the `espidf` one correctly, but I'm guessing there are other embedded-first architectures that would want something similar.

---

Check out these sweet-:peach: results!

```
$ valgrind --tool=dhat target/dr/ftzz g -n 1M /tmp/.tmp$RANDOM
```

**Old:**

```
Total:     437,291 bytes (100%, 2,288.6/Minstr) in 4,275 blocks (100%, 22.37/Minstr), avg size 102.29 bytes, avg lifetime 27,154,908.14 instrs (14.21% of program duration)
```

**New:**

```
Total:     335,808 bytes (100%, 1,761.28/Minstr) in 1,864 blocks (100%, 9.78/Minstr), avg size 180.15 bytes, avg lifetime 61,395,245.76 instrs (32.2% of program duration)
```