@SUPERCILEX 2022-04-21
It seems like InvalidInput is expected: https://github.com/rust-lang/rust/blob/09ccb6c59d026b94edae50ba949b22dfc5d65ed1/src/test/ui/paths-containing-nul.rs#L15 Is this something that can change or nah?