@SUPERCILEX 2022-04-21
F

Can someone run that for me? The compiler is I/O heavy, so I'd be curious to see if this did anything.