@SUPERCILEX 2022-04-21
@RalfJung I've been looking at the assembly of `[MaybeUninit::uninit(); N]` and I'm pretty sure that ends up memsetting the array whereas `MaybeUninit::<[u8; N]>::uninit()` doesn't. That seems like a serious bug? Based on these docs, the two should be equivalent: https://doc.rust-lang.org/stable/std/mem/union.MaybeUninit.html#method.uninit_array.

Godbolts: https://godbolt.org/z/341fEjqMd, https://godbolt.org/z/K9dMx8Whr