@SUPERCILEX 2022-04-21
Similarly, who's in charge LLVM intrinsics? `unlikely` just straight up doesn't work. :( https://godbolt.org/z/aMaTzEr3q The hot path code ends up last.