@Kobzol 2022-04-21
If you want to separate cold/hot code, it's better to put the cold code into a function (it can be a local function) marked with `#[inline(never)]`. There is also a `#[cold]` attribute, but that didn't worked out that great for me by itself. I suspect using this "trick" here will be a bit annoying though.

If you look at the generated [LLVM IR](https://play.rust-lang.org/?version=nightly&mode=debug&edition=2021&gist=e885f288f23ec96ec7a8f4605be852cc) (Build button at top left -> LLVM IR), you can see that `rustc` correctly emits `llvm.expect` [intrinsics](https://llvm.org/docs/LangRef.html#llvm-expect-intrinsic). So this is more of a "fault" of LLVM than Rust. Sadly I suppose that there is no guarantee that the compiler will actually respect these annotations. It might have decided that this order is for the best. Rust uses PGO for dist builds though, so that might help in this case.