@Kobzol 2022-04-21
Good catch with the uninit array. As a workaround, if you use `let mut p: [MaybeUninit<u8>; 571] = MaybeUninit::uninit_array();`, it doesn't call `memset`.