@SUPERCILEX 2022-04-21
@Kobzol Yeah, that's what I ended up doing. Something super interesting is that those `.expects` don't show up in release mode: https://play.rust-lang.org/?version=nightly&mode=release&edition=2021&gist=0b937a6d7173a5bdf34bd5197ab5ca44. Could that be the bug? That'd be pretty scary if so. I checked on godbolt and debug mode does seem better behaved order wise (though there's so much going on it's hard to tell).

As for the uninitialized memory, I really want that to be a bug. :) Otherwise my world has been flipped upside down and I have to go fix a bunch of other code that assumed uninitialized memory actually meant uninitialized. I mean, the docs literally say uninit_array will be removed because const expressions will replace it. That's a bit of a painful lie if it's not true.