@Kobzol 2022-04-21
Yeah that sounds like a bug. Would you be willing to open an issue with a minimal reproducible example? If not, I can do it later.
`llvm.expect` disappearing also sounds more like a bug than a feature.