@Kobzol 2022-04-21
Yeah it outright seems that release strips the `expect` annotations. I'll file an issue about that.