@SUPERCILEX 2022-10-09
Dang, I'm really not quite sure what to make of these benchmarks. They seem to be flip flopping. My own testing shows a ~4x improvement in instruction count and very minor improvement in real-time in an isolated micro-benchmark.

## Test code

```rust
    for _ in 0..10_000_000 {
        let z = File::open("Cargo.toml").unwrap();
        std::hint::black_box(&z);
    }
```

## Hyperfine

```
Benchmark 1: ./old
  Time (mean ± σ):     11.756 s ±  0.226 s    [User: 1.270 s, System: 10.418 s]
  Range (min … max):   11.562 s … 12.268 s    10 runs
 
Benchmark 2: ./new
  Time (mean ± σ):     11.232 s ±  0.153 s    [User: 0.884 s, System: 10.274 s]
  Range (min … max):   11.076 s … 11.634 s    10 runs
 
Summary
  './new' ran
    1.05 ± 0.02 times faster than './old'
```

## Old

```
==62778== DHAT, a dynamic heap analysis tool
==62778== Copyright (C) 2010-2018, and GNU GPL'd, by Mozilla Foundation
==62778== Using Valgrind-3.18.1 and LibVEX; rerun with -h for copyright info
==62778== Command: ../target/release/rmz
==62778== 
==62778== 
==62778== Total:     110,002,157 bytes in 10,000,010 blocks
==62778== At t-gmax: 1,736 bytes in 3 blocks
==62778== At t-end:  0 bytes in 0 blocks
==62778== Reads:     90,019,054 bytes
==62778== Writes:    120,011,612 bytes
==62778== 
==62778== To view the resulting profile, open
==62778==   file:///usr/libexec/valgrind/dh_view.html
==62778== in a web browser, click on "Load...", and then select the file
==62778==   /home/asaveau/Desktop/fuc/rmz/dhat.out.62778
==62778== The text at the bottom explains the abbreviations used in the output.
```

```
Reading data from 'callgrind.out.65337'...
--------------------------------------------------------------------------------
Profile data file 'callgrind.out.65337' (creator: callgrind-3.18.1)
--------------------------------------------------------------------------------
I1 cache: 32768 B, 64 B, 8-way associative
D1 cache: 32768 B, 64 B, 8-way associative
LL cache: 16777216 B, 64 B, 16-way associative
Timerange: Basic block 0 - 1030069978
Trigger: Program termination
Profiled target:  ../target/dr/rmz (PID 65337, part 1)
Events recorded:  Ir Dr Dw I1mr D1mr D1mw ILmr DLmr DLmw Bc Bcm Bi Bim
Events shown:     Ir Dr Dw I1mr D1mr D1mw ILmr DLmr DLmw Bc Bcm Bi Bim
Event sort order: Ir Dr Dw I1mr D1mr D1mw ILmr DLmr DLmw Bc Bcm Bi Bim
Thresholds:       99 0 0 0 0 0 0 0 0 0 0 0 0
Include dirs:     
User annotated:   
Auto-annotation:  on

--------------------------------------------------------------------------------
Ir                     Dr                     Dw                   I1mr           D1mr           D1mw         ILmr           DLmr           DLmw         Bc                   Bcm                 Bi                  Bim          
--------------------------------------------------------------------------------
4,300,336,369 (100.0%) 1,080,081,361 (100.0%) 730,035,268 (100.0%) 1,679 (100.0%) 2,329 (100.0%) 854 (100.0%) 1,615 (100.0%) 1,807 (100.0%) 793 (100.0%) 740,060,749 (100.0%) 10,004,883 (100.0%) 50,000,898 (100.0%) 383 (100.0%)  PROGRAM TOTALS

--------------------------------------------------------------------------------
Ir                   Dr                   Dw                   I1mr        D1mr         D1mw         ILmr        DLmr         DLmw         Bc                   Bcm                 Bi                  Bim          file:function
--------------------------------------------------------------------------------
640,000,640 (14.88%) 200,000,200 (18.52%) 110,000,110 (15.07%)  7 ( 0.42%)   0            0           7 ( 0.43%)   0            0           90,000,090 (12.16%)          3 ( 0.00%)          .           .           ./malloc/./malloc/malloc.c:_int_free [/usr/lib/x86_64-linux-gnu/libc.so.6]
580,000,000 (13.49%) 100,000,000 ( 9.26%)           0           1 ( 0.06%)   0            0           1 ( 0.06%)   0            0          220,000,000 (29.73%) 10,000,008 (99.95%)          .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/core/src/slice/memchr.rs:<&[u8] as alloc::ffi::c_str::CString::new::SpecNewImpl>::spec_new_impl
430,000,313 (10.00%) 110,000,080 (10.18%)  60,000,027 ( 8.22%)  8 ( 0.48%)   1 ( 0.04%)   0           7 ( 0.43%)   1 ( 0.06%)   0           70,000,067 ( 9.46%)         32 ( 0.00%)          .           .           ./malloc/./malloc/malloc.c:malloc [/usr/lib/x86_64-linux-gnu/libc.so.6]
390,000,000 ( 9.07%) 130,000,000 (12.04%)  20,000,000 ( 2.74%)  5 ( 0.30%)   0            0           5 ( 0.31%)   0            0           90,000,000 (12.16%)         10 ( 0.00%) 10,000,000 (20.00%)  1 ( 0.26%)  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/std/src/sys/unix/fs.rs:std::fs::OpenOptions::_open
340,000,000 ( 7.91%)  70,000,000 ( 6.48%)  40,000,000 ( 5.48%)  3 ( 0.18%)   0            0           3 ( 0.19%)   0            0           50,000,000 ( 6.76%)          .                   .           .           ./io/../sysdeps/unix/sysv/linux/open64.c:open [/usr/lib/x86_64-linux-gnu/libc.so.6]
290,000,000 ( 6.74%)  80,000,000 ( 7.41%) 140,000,000 (19.18%)  3 ( 0.18%)   0            0           3 ( 0.19%)   0            0           20,000,000 ( 2.70%)          .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/ffi/c_str.rs:<&[u8] as alloc::ffi::c_str::CString::new::SpecNewImpl>::spec_new_impl [/home/asaveau/Desktop/fuc/target/dr/rmz]
220,000,198 ( 5.12%)  80,000,072 ( 7.41%)  40,000,036 ( 5.48%)  3 ( 0.18%)   0            0           3 ( 0.19%)   0            0           30,000,027 ( 4.05%)          6 ( 0.00%)          .           .           ./malloc/./malloc/malloc.c:free [/usr/lib/x86_64-linux-gnu/libc.so.6]
180,000,000 ( 4.19%)  70,000,000 ( 6.48%)  60,000,000 ( 8.22%)  2 ( 0.12%)   0            0           2 ( 0.12%)   .            .                    .                   .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/std/src/fs.rs:std::fs::OpenOptions::_open [/home/asaveau/Desktop/fuc/target/dr/rmz]
150,000,015 ( 3.49%)  60,000,006 ( 5.56%)  50,000,005 ( 6.85%)  2 ( 0.12%)   0            0           2 ( 0.12%)   .            .                    .                   .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/ffi/c_str.rs:alloc::ffi::c_str::CString::_from_vec_unchecked [/home/asaveau/Desktop/fuc/target/dr/rmz]
130,000,960 ( 3.02%)  30,000,255 ( 2.78%)  20,000,170 ( 2.74%)  4 ( 0.24%)   1 ( 0.04%)   3 ( 0.35%)  4 ( 0.25%)   1 ( 0.06%)   3 ( 0.38%)  30,000,204 ( 4.05%)         25 ( 0.00%)          .           .           ./string/../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S:__memcpy_avx_unaligned_erms [/usr/lib/x86_64-linux-gnu/libc.so.6]
120,000,012 ( 2.79%)  40,000,004 ( 3.70%)  20,000,002 ( 2.74%)  2 ( 0.12%)   0            0           2 ( 0.12%)   0            0           20,000,002 ( 2.70%)          6 ( 0.00%)          .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/vec/mod.rs:alloc::ffi::c_str::CString::_from_vec_unchecked
 90,000,000 ( 2.09%)  20,000,000 ( 1.85%)           0           1 ( 0.06%)   0            0           1 ( 0.06%)   0            0           20,000,000 ( 2.70%)          .                   .           .           ./io/../sysdeps/unix/sysv/linux/close.c:close [/usr/lib/x86_64-linux-gnu/libc.so.6]
 90,000,000 ( 2.09%)  10,000,000 ( 0.93%)           0           2 ( 0.12%)   0            0           2 ( 0.12%)   0            0           10,000,000 ( 1.35%)          .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/std/src/sys/unix/mod.rs:std::fs::OpenOptions::_open
 70,000,014 ( 1.63%)  20,000,004 ( 1.85%)  10,000,002 ( 1.37%)  1 ( 0.06%)   1 ( 0.04%)   0           1 ( 0.06%)   0            0           10,000,002 ( 1.35%)          5 ( 0.00%) 10,000,002 (20.00%)  1 ( 0.26%)  ???:__rust_alloc [/home/asaveau/Desktop/fuc/target/dr/rmz]
 70,000,000 ( 1.63%)           0           20,000,000 ( 2.74%)  .            .            .           .            .            .                    .                   .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/std/src/fs.rs:rmz::main
 70,000,000 ( 1.63%)           0                    0           1 ( 0.06%)   0            0           1 ( 0.06%)   0            0           20,000,000 ( 2.70%)          1 ( 0.00%)          .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/raw_vec.rs:<&[u8] as alloc::ffi::c_str::CString::new::SpecNewImpl>::spec_new_impl
 60,000,000 ( 1.40%)           0           20,000,000 ( 2.74%)  0            0            0           0            0            0           10,000,000 ( 1.35%)          .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/core/src/result.rs:std::fs::OpenOptions::_open
 50,000,000 ( 1.16%)  10,000,000 ( 0.93%)  20,000,000 ( 2.74%)  0            0            0           0            0            0                    0                   0          10,000,000 (20.00%)  1 ( 0.26%)  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/vec/mod.rs:<&[u8] as alloc::ffi::c_str::CString::new::SpecNewImpl>::spec_new_impl
 40,000,000 ( 0.93%)           0           20,000,000 ( 2.74%)  .            .            .           .            .            .                    .                   .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/ffi/c_str.rs:std::fs::OpenOptions::_open
 40,000,000 ( 0.93%)           0           10,000,000 ( 1.37%)  0            0            0           0            0            0           10,000,000 ( 1.35%)          .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/alloc.rs:std::fs::OpenOptions::_open
 30,000,027 ( 0.70%)           0                    0           0            0            0           0            0            0           10,000,009 ( 1.35%)          7 ( 0.00%)          .           .           ./malloc/./malloc/arena.c:free
 30,000,022 ( 0.70%)  10,000,008 ( 0.93%)  20,000,006 ( 2.74%)  1 ( 0.06%)   0            0           1 ( 0.06%)   .            .                    .                   .                   .           .           src/main.rs:rmz::main [/home/asaveau/Desktop/fuc/target/dr/rmz]
 30,000,003 ( 0.70%)  10,000,001 ( 0.93%)           0           0            0            0           0            0            0           10,000,001 ( 1.35%)          3 ( 0.00%)          .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/raw_vec.rs:alloc::ffi::c_str::CString::_from_vec_unchecked
 30,000,000 ( 0.70%)  20,000,000 ( 1.85%)           0           1 ( 0.06%)   0            0           1 ( 0.06%)   0            0           10,000,000 ( 1.35%)          .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/core/src/result.rs:rmz::main
 30,000,000 ( 0.70%)           0           30,000,000 ( 4.11%)  1 ( 0.06%)   0            0           1 ( 0.06%)   .            .                    .                   .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/std/src/sys/unix/fs.rs:rmz::main
 30,000,000 ( 0.70%)           0           10,000,000 ( 1.37%)  .            .            .           .            .            .                    .                   .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/alloc.rs:<&[u8] as alloc::ffi::c_str::CString::new::SpecNewImpl>::spec_new_impl
 20,000,000 ( 0.47%)           .                    .           .            .            .           .            .            .                    .                   .                   .           .           /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/core/src/num/uint_macros.rs:<&[u8] as alloc::ffi::c_str::CString::new::SpecNewImpl>::spec_new_impl
 10,000,002 ( 0.23%)  10,000,002 ( 0.93%)           0           1 ( 0.06%)   0            0           1 ( 0.06%)   0            0                    0                   0          10,000,002 (20.00%)  1 ( 0.26%)  ???:__rust_dealloc [/home/asaveau/Desktop/fuc/target/dr/rmz]

--------------------------------------------------------------------------------
-- Auto-annotated source: src/main.rs
--------------------------------------------------------------------------------
Ir                  Dr                  Dw                  I1mr       D1mr       D1mw       ILmr       DLmr DLmw Bc Bcm Bi Bim 

-- line 27 ----------------------------------------
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               #[clap(flatten)]
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               verbose: Verbosity,
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               #[arg(short, long, short_alias = '?', global = true)]
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               #[arg(action = ArgAction::Help, help = "Print help information (use `--help` for more detail)")]
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               #[arg(long_help = "Print help information (use `-h` for a summary)")]
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               help: Option<bool>,
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           }
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           
        14 ( 0.00%)          1 ( 0.00%)          6 ( 0.00%) 1 ( 0.06%) 0          0          1 ( 0.06%) .          .          .          .          .          .           fn main() {
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               // let args = Rmz::parse();
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               //
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               // RemoveOp::builder()
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               //     .files(args.files.iter().map(AsRef::as_ref))
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               //     .build()
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               //     .run()
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               //     .unwrap();
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               for _ in 0..10_000_000 {
10,000,000 ( 0.23%)          0          10,000,000 ( 1.37%) .          .          .          .          .          .          .          .          .          .                   let z = File::open("Cargo.toml").unwrap();
10,000,000 ( 0.23%)          0          10,000,000 ( 1.37%) .          .          .          .          .          .          .          .          .          .                   std::hint::black_box(&z);
10,000,000 ( 0.23%) 10,000,000 ( 0.93%)          .          .          .          .          .          .          .          .          .          .          .               }
         8 ( 0.00%)          7 ( 0.00%)          .          .          .          .          .          .          .          .          .          .          .           }
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           #[cfg(test)]
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           mod cli_tests {
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               use std::io::Write;
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               use clap::{Command, CommandFactory};
         .                   .                   .          .          .          .          .          .          .          .          .          .          .               use goldenfile::Mint;
         .                   .                   .          .          .          .          .          .          .          .          .          .          .           
-- line 56 ----------------------------------------

--------------------------------------------------------------------------------
The following files chosen for auto-annotation could not be found:
--------------------------------------------------------------------------------
  ./io/../sysdeps/unix/sysv/linux/close.c
  ./io/../sysdeps/unix/sysv/linux/open64.c
  ./malloc/./malloc/arena.c
  ./malloc/./malloc/malloc.c
  ./string/../sysdeps/x86_64/multiarch/memmove-vec-unaligned-erms.S
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/alloc.rs
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/ffi/c_str.rs
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/raw_vec.rs
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/alloc/src/vec/mod.rs
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/core/src/num/uint_macros.rs
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/core/src/result.rs
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/core/src/slice/memchr.rs
  /rustc/f83e0266cf7aaa4b41505c49a5fd9c2363166522/library/std/src/fs.rs
…